Implement CRC32C hashes for ObjectInsert().
With this change the library sends (unless explicitly disabled by the
application developer) the CRC32C checksum of any data included in
InsertObjectMedia() calls. The integration tests in this change verify
that valid CRC32C checksums are accepted, and that invalid CRC32C
checksums are rejected by the testbench (and the service).