Add option to force resumable uploads.
The library automatically selects resumable uploads when it can
determine that this is the best upload strategy. But sometimes the
application simply knows that a resumable upload is necessary, in this
case the application developer should be able to force a resumable
upload. In addition, setting this option to a non-empty string will be
used to restore a previously created resumable upload.