Idempotency policies for conditional mutations.
This fixes #1390.

Conditional mutations are sometimes idempotent, e.g. setting a value in
one column if another satisfies a condition. This PR adds an overridable
member function to IdempotentMutationPolicy to allow usres to define
custom behavior.

The general solution needs much more design a work, hence #1715 was
created to address it.