Do not throw exceptions from ObjectWriteStream.
Unless the application explicitly requests exceptions using
std::ostream::exceptions(). By default std::ios_base::badbit is set when
an error is detected, and the ObjectWriteStream saves the status with a
more detailed error message.

This change will reduce code coverage because many errors that used to
be propagated as exceptions now require an extra if(), but we do not
have test infrastructure to produce these error conditions on demand.