Change *Notification() functions to return StatusOr.
In this change all the `storage::Client::*Notification()` functions are
changed to return `StatusOr<>`. This change also introduces new helpers
to write unit tests for functions that return `StatusOr<>` instead of
throwing exceptions.

The examples were changed to simply call `.value()`, that keeps them
free of error checking noise, and still produces good error messages if
there is ever an error.

The integration tests are changed to check the if the returned
`StatusOr<>` is `.ok()`, and then we use the `operator*()` and
`operator->()` to use the `StatusOr<T>` as a pointer to `T`.