Fix build problems with GCC-7.3.
The build for Ubuntu:18.04 breaks with a optional<StatusOr<T>>. It
probably should be fixed, we must be doing something too complicated. On
the other hand, we do not need an `optional<StatusOr<T>>`, a "plain"
`StatusOr<T>` is flexible enough to store a `T` or a "you tried to read
an invalid iterator", so do that instead.