Only use StatusCode values in Status.
Remove the constructors that accept `long`, and only use the values in
the StatusCode enum. We use the `storage::AsStatus()` function to map
the HttpResponse to the correct `Status`.

Lots of changes in tests because instead of checking for HTTP codes we
need to check for the corresponding StatusCode.

The more subtle change is what happens with `IsRetryable`, I kept the
semantics of its tests unchanged, so the behavior should be the same.