Add trace messages to troubleshoot flake.
The TableReadMultipleCellsBigValue test fails (rarely) with a timeout,
but there is no indication of which RPC timed out. This PR adds a number
of SCOPED_TRACE statements to get a better idea of what fails on a
failure.

In addition, the test fine tunes the message size. The intention of the
test is to verify that we can write and read cells much larger than the
default gRPC maximum message size (~4MiB).

This default is too small for Cloud Bigtable, where cells can be much
larger than 4MiB. The library achieves this by setting a larger value
(~256MiB) for the maximum message size.

The test was creating a row of ~250MiB. I think testing with 128MiB is
more reasonable. Large enough to prove that we are not using the 4MiB
setting, yet small enough to finish quickly and avoid trouble with any
message overhead.

I was able to run the test with the new setting about 1000 times without
problem. But note that other issues (contention in the bigtable master,
quota limits on the number of tables created by hour / day / etc.) could
also turn this test flaky. So I am leaving the debug messages on for
now.