Create a single table per integration test suite.
We used to create a separate table for each test case in the integration
tests. That makes the test case easier to write (and reason about), but
creating a table is very slow in the production tests, and one can
easily hit the daily or peak quota for Bigtable:

https://cloud.google.com/bigtable/quotas#operation-quotas

In this PR we change the integration tests to share a common table. I
had to normalize the table schema: typically we had column families
named "family{1,2,3}" but sometimes it was "fam{0,1,2,3}". I also
changed the SetUp() function to drop all the rows in the table, so at
least the table is clean.

Note that some of the instance admin and table admin tests still create
tables, but those are typically disabled when running against
production.