bug: Account for when CreateHashValidator's options are set to false
Fixes #2978

Previously, the `CreateHashValidator` functions did not check the value of the
options, only the existence of the options. As a consequence, whether the user
specifies `true` or `false` doesn't matter. This PR checks the values of both
options and adds the logic to both overloads. One hitch (to preserve existing
behavior) is that the options are mutually exclusive IFF only one option is
specified.  It's possible that this approach can be simplified, but boolean
logic is hard.