feat: move creating prefix marker to ComposeMany
This is needed for #3279.

Before this PR, the user had to create a prefix (which could fail), pass
it to `ComposeMany` and remove it afterwards. It was a lot of boilerplate.

With this MR, the prefix name is picked at random and `ComposeMany`
handles creating and deleting the marker file.