feat: implement resumable parallel fule uploads
This fixes #2951.

This bases on the lower-level implementation from #3373. The API is not
changed, but the user can now specify `UseResumableSession` option.

Under the hood, the implementation piggy-backs file-related information
to persistent state (i.e. how the file was split into streams) in order
to allow for changes in how the files are split.