fix: use network to check if running inside Google
This fixes #3304

In order to tell if the code is running inside Google (GCE, GKE, Cloud
Run, AppEngine flexible) we now check if `metadata.google.internal`
resolves to an IP (v4 od v6). Any other attempts failed due to gVisor
(used in Cloud Run) isolating the host operating system very well.

If network is disabled or customized when run on Cloud Run, this won't
work, however it's and unlikely combination.