fix: deadlock in MutationBatcher
This fixes #4083.

The used to happen when the continuation of `AsyncBulkApply` was
executed synchronously (which happened if the operation finished really
quickly). In order to overcome it, the continuation now schedules the
continuation on the completion queue explicitly. A solution in which
`FlushIfPossible` was releasing the mutex was hard because the mutex
would have to be reacquired later making the code very brittle and
perhaps incorrect.