fix(storage)!: incorrect type in OLM condition
**BREAKING CHANGE**: changed the `created_before` field in
`LifecycleRuleCondition` from a time point to a date. The service
rejects timestamps, it only accepts dates in YYYY-mm-dd format. We
considered trying to keep the C++ type the same and fixing the
conversions to the underlying protocol, but (a) this was confusing to
any users, and (b) we do not expect anybody was using this field because
it simply did not work before.