refactor: implement spanner::Timestamp using absl::Time
This PR removes lots of time logic that is no longer needed since we can
use `absl::Time`. However, there are some semantic changes that should
be discussed:

1. Should spanner::Timestamp *only* support times in its documented
range? That is, should it allow times outside that range that happen to
fit due to our implementation?

2. Should spanner::Timestamp::get allow extracting sys_time<pico> or
anything smaller than nanos? Or rather, *must* this be supported?