feat(pubsub): reduce contention in lease management
I found that the lease management layer suffered from too much
contention. Basically each `ack()` and `nack()` call required a lock, as
the hash map with all the pending messages is updated. After updating
this hash map the code calls the `StreamingSubscriptionBatchSource`
stage to do more work, which requires additional locking. This PR
collapses the functionality of the `SubscriptionLeaseManagement` stage
onto `StreamingSubscriptionBatchSource`, reducing the number of locks
per `ack()`/`nack()` call.

These improvements in locking overhead are minor, but collapsing the two
layers simplifies other changes I want to make, such as (1) suspend
`Read()` calls if the number (and/or size) of outstanding messages is
too high, (2) create multiple AsyncPull streams on the same session, and
(3) maybe avoid copying the large `ack_id` strings (160 bytes!) and use
a more efficient in-memory representation.

I also simplified the timer management to periodically update the
leases. We use a single periodic timer to check if the leases need an
update, and piggy back the update with any other pending nacks and/or
acks. That is less accurate, but the periodic timer is executed about
100 times a second, and the leases need to be updated every 10 seconds
(we actually do it ever 5). Good enough I think.