feat: remove `CompletionQueue` from InstanceAdmin
This is a part of #2567.

This PR removes the `CompletionQueue` parameter from the `InstanceAdmin`
API. Apart from the obvious changes, additional option to
`bigtable::ConnectioOptions` was added to allow for custom worker
threads.

`AutomaticallyCreatedBackgroundThreads` now cancels all outstanding
operations in the destructor. Without this, destroying this object after
using it in clients with client refresh was taking multiple minutes.

A non-trivial concurrency change in the `CompletionQueue` was necessary.
Before this PR the `StartOp` passed to `StartOperation` was called with
the `CompletionQueue` lock held. This became problematic with this
change because the same completion queue is used for connection refresh
and all user activity. User activity, apart from calling the appropriate
method on the stub was also scheduling timers, yielding a deadlock.