cleanup(storage): consistent CMake target names
Export the storage client library as `google-cloud-cpp::storage`. The
legacy name for this library (`pubsub_client` is also exported for
backwards compatibility. The legacy name generates warnings with
CMake >= 3.18.

I also created an alias and exported name for the GCS+gRPC library. This
library is *not* installed in a normal build, so its name is not that
important. Nevertheless, I called it
`google-cloud-cpp::experimental-gcs-grpc` to make clear that it is not
ready for production.