cleanup: internal::MakeOptions accepts by value
A value (copy) is needed because we move the fields from the old
`ConnectionOptions` to the new `Options`. Previously it was requiring an
rvalue reference, which annoyingly required the caller to have to make
an explicit copy sometimes (e.g., if the caller only had a const-ref and
couldn't move).