ci(gdb): enable GCS tests that require json and p12 keyfiles
I've stored the base64-encoded data for these files in the Secret
Manager. The Secrete Manager will inject env vars into our build step
contianing the base64-encoded data for each file. The test will then
decode that data and write it to a file for use in the test.

When we rotate key files, we will need to update the version of the
secret in Secret Manager.