ci: build with older versions of each dep
This build compiles with older versions of each dependency. Its only
purpose is to verify compilation (and basic testing), so we only run
this in one platform. I used vcpkg to fetch the older versions of the
deps, including their dependencies at the time.

The only magic number is the `vcpkg` baseline, which is obtained by
looking at the history of:

https://github.com/microsoft/vcpkg/blob/master/versions/baseline.json

and finding the commit that has all the versions we want. In this build
vcpkg reports:

```
    abseil[core]:x64-linux -> 2020-09-23#3
    benchmark[core]:x64-linux -> 1.5.2
    c-ares[core]:x64-linux -> 1.17.1
    crc32c[core]:x64-linux -> 1.1.1
    curl[core,non-http,openssl,ssl]:x64-linux -> 7.74.0#3
    grpc[core]:x64-linux -> 1.33.1#3
    gtest[core]:x64-linux -> 1.10.0#4
    nlohmann-json[core]:x64-linux -> 3.9.1
    openssl[core]:x64-linux -> 1.1.1j
    protobuf[core]:x64-linux -> 3.14.0#3
    re2[core]:x64-linux -> 2020-10-01
    upb[core]:x64-linux -> 2020-08-19
    zlib[core]:x64-linux -> 1.2.11#9
```