mock_response_reader: prevent behavior change from a gRPC change.
This class currently encourages an ownership hack that works around a gRPC API
issue that is [being fixed](https://github.com/grpc/proposal/pull/238). At that
point the tests that use this class will start failing, because they manage it
using **two different `std::unique_ptr` objects**.

Prevent them failing by overriding the upcoming `Destroy` method to do nothing.
This preserves the existing behavior. Because I haven't used the `override`
keyword, this works before and after the gRPC change, decoupling this issue
from the import of the gRPC change.

Longer term it would be better to make the use of this class not require the
ownership hack at all, but that is a larger project. In the meantime this
change unblocks fixing the need for such hacks for other subclasses that don't
want to further promulgate the issue.