fix(storage): alternative endpoints and Host header
Some alternatives for `storage.googleapis.com` require setting the
`Host:` header to `storage.googleapis.com`. This change (I think)
simplifies the host-header computation: we let `libcurl` do its thing in
most cases (it pulls it from the URL), and we override it only in the
cases where we know it is needed.