cleanup(storage): simplify integration test cleanup
Most of our integration tests create some objects, which then need to be
deleted when the test ends. In addition to having to remember to delete
the object, some of the tests are a bit unnatural, as they should use
`ASSERT_*()` but they use `EXPECT_*()` to ensure the cleanup step runs.

With this change the cleanup is performed by a common `TearDown()`
function, the tests just add objects (and buckets if applicable) to a
list that will be deleted at the end.

In some tests the `DeleteObject()` function is actually part of the
test, i.e., we are testing the `DeleteObject()` function. In these cases
I left the code alone. I also left unchanged some tests where objects
are created without a `gcs::ObjectMetadata` becoming available,
typically because the test is using an streaming upload  implicitly
finalized by the stream's destructor.